% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2_asymptotic_test.R
\name{compute_chi2_power}
\alias{compute_chi2_power}
\title{Chi2 based power analysis.}
\usage{
compute_chi2_power(
  n_cases,
  n_controls,
  maf_snp1,
  maf_snp2,
  a_OR,
  b_OR,
  int_OR,
  a1,
  a2,
  b1,
  b2,
  i,
  prevalence,
  penetrance_ref_genotype,
  alpha,
  encoding = "dosage",
  input_type
)
}
\arguments{
\item{n_cases}{Number of cases}

\item{n_controls}{Number of controls}

\item{maf_snp1}{maf of the snp1}

\item{maf_snp2}{maf of the snp2}

\item{a_OR}{Odds Ratio for the additive effect of snp1}

\item{b_OR}{Odds Ratio for the additive effect of snp2}

\item{int_OR}{Odds Ratio for the interaction effect of snp1 and snp2}

\item{prevalence}{Prevalence of the disease}

\item{penetrance_ref_genotype}{Penetrance of the reference genotype}

\item{alpha}{the alpha threshold for overall significance}

\item{encoding}{the encoding of snp1 and snp2. Possibilities are "dosage","dominant","recessive","hetero","dosage_recessive","dominant_"recessive","hetero_recessive".}

\item{input_type}{the type of input; either "OR" for Odds ratio or "GRR" for Genotypic Risk Ratio}
}
\value{
a sinlge float. The power
}
\description{
Genotypes frequencies are computed separately for cases and
controls based on the prevalence of the diseases, the minor allele frequency of the 2 snps and their ORs.
n_cases cases and n_controls are then simulated using the genotype frequencies and used in a logistic regression
model to predict the diseases phenotype. Under the null hypothesis, the Wald test statistics of the coefficients follows a Chi-square distribution with 1df.
Under the alternative hypothesis, the Wald test statistic for the coefficients
follow a chi-square distribution with a non-centrality parameter equal to its expectation. The power is computed as the cumulative distribution of the
Wald test-static under the alternative hypothesis up to the 1-alpha quantile of the Wald-test statistic under the null.
}
\examples{

compute_chi2_power(n_cases = 1000,n_controls = 1000,maf_snp1 = 0.1,maf_snp2 = 0.1,
a_OR = 1,b_OR = 2,int_OR = 1.2,penetrance_ref_genotype = 0.1,
alpha = 1e-5,encoding = "dosage",input_type="OR")
}
