% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epilogitpower.R
\docType{package}
\name{epilogitpower}
\alias{epilogitpower}
\title{epilogitpower: A package for power analysis in epistasis studies}
\description{
The epilogitpower provides functions to compute the power of logistic regression
based epistasis studies dependent on the effect sizes and minor allele frequency (MAF)
of the SNPs as well as the prevalence of the disease.
}
\section{compute_chi2_power }{

compute power based on the chi-square distribution of coefficients
}

\section{MC_power_analysis}{

compute power using Monte-Carle Simultations
}

