% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_simulations.R
\name{MC_power_analysis}
\alias{MC_power_analysis}
\title{Monte Carlo based power analysis}
\usage{
MC_power_analysis(
  n_cases,
  n_controls,
  maf_snp1,
  maf_snp2,
  a_OR = NULL,
  b_OR = NULL,
  int_OR = NULL,
  a1 = NULL,
  a2 = NULL,
  b1 = NULL,
  b2 = NULL,
  i = NULL,
  prevalence = NULL,
  penetrance_ref_genotype = NULL,
  encoding = "dosage",
  input_type,
  alpha,
  n_simulations = 100,
  threads = 1
)
}
\arguments{
\item{n_cases}{the number of cases to sample if sampling scheme is "stratified"}

\item{n_controls}{the number of controls to sample if sampling scheme is "stratified"}

\item{maf_snp1}{maf of the snp1}

\item{maf_snp2}{maf of the snp2}

\item{a_OR}{Odds ratio for the additive effect of snp1}

\item{b_OR}{Odds ratio for the additive effect of snp2}

\item{int_OR}{Odds ratio for the interaction effect of snp1 and snp2}

\item{prevalence}{prevalence of the disease}

\item{encoding}{the genetic model encoding to choose from "dosage","dominant","recessive","hetero" with respect to minor allele. Only}

\item{alpha}{the alpha threshold for overall significance}

\item{n_simulations}{number of Monte Carlo simulations to run}

\item{threads}{number of threads to use for the MC simulations}

\item{sample_size}{the total sample size if the sampling scheme is "naive"}

\item{sampling_method}{method used to generate the dataset. Either "naive" randomly drawing from a common population determined by the mafs or "Stratified" drawing cases and controls separately}
}
\value{
a sinlge float. The power
}
\description{
Genotypes frequencies are computed separately for cases and
controls based on the prevalence of the diseases, the minor allele frequency of the 2 snps and their ORs.
n_cases cases and n_controls are then simulated using the genotype frequencies and used in a logistic regression
model to predict the diseases phenotype. The power is then computed by estimating the distribution of the Wald-test statistics under the alternative
by Monte-Carlo simulations.
}
\examples{
MC_power_analysis(n_controls=1000,n_cases=1000,maf_snp1 = 0.1,maf_snp2 = 0.1,a1=1.2,a2=0.9,b1=1.14,b2=1.02,i=1.23,prevalence = 0.1,encoding="dosage",input_type="GRR",alpha = 0.05,n_simulations = 10)
MC_power_analysis(n_controls=1000,n_cases=1000,maf_snp1 = 0.1,maf_snp2 = 0.1,a_OR=0.9,b_OR=1.1,int_OR = 1.5,prevalence = 0.1,encoding="dosage",penetrance_ref_genotype=0.09,input_type="OR",alpha = 0.05,n_simulations = 10)
}
