% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_ORs.R
\name{possible_int_OR}
\alias{possible_int_OR}
\title{Prediciton of possible interaction OR.}
\usage{
possible_int_OR(
  penetrance_ref_genotype = NULL,
  prevalence_up = 1,
  prevalence_low = 0,
  int_OR_upper = 2,
  int_OR_lower = 0.5,
  mafa,
  mafb,
  a_OR,
  b_OR,
  encoding = "dosage",
  plot = FALSE
)
}
\arguments{
\item{penetrance_ref_genotype}{The prevalence of the reference genotype}

\item{prevalence_up}{The upper bound of the range of possible prevalences. Default=1}

\item{prevalence_low}{The lower bound of the range of possible prevalences. Default=0}

\item{int_OR_upper}{The upper bound of the range of possible ORs for the interaction. Default=2}

\item{int_OR_lower}{The lower bound of the range of possible ORs for the interaction. Default=0.5}

\item{mafa}{maf of the snp1}

\item{mafb}{maf of the snp2}

\item{a_OR}{Odds Ratio for the additive effect of snp1}

\item{b_OR}{Odds Ratio for the additive effect of snp2}

\item{encoding}{Genetic encoding models for the two snps}

\item{plot}{Wheter to return a plot of the predicted prevalence in function of the interaction OR.}
}
\value{
a list of the maximum and minimum OR compatible with the upper and lower bound of the prevalence. If plot=T also returns a plot
of the predicted prevalence in function of the interaction OR.
}
\description{
Predicts the range of possible ORs for the interaction given data about the marginal effect of both SNPs
and epidemiological data.
}
\examples{
possible_int_OR(penetrance_ref_genotype = 0.0866,mafa = 0.22,mafb = 0.015,a_OR = 1.03,b_OR = 1.46,plot = FALSE,int_OR_upper = 2)
possible_int_OR(penetrance_ref_genotype = 0.0866,mafa = 0.22,mafb = 0.015,a_OR = 1.03,b_OR = 1.46,plot = TRUE,int_OR_upper = 2)
}
